Refactor: renaming, and logic export
* old useScooter pseudonym -> useMobility
* routed import/export of new useOperators through ...logic/index.ts
