+++ src/containers/Admin/EditTab/ScooterPanel/index.tsx
@@ -5,7 +5,7 @@ import { Label } from '@entur/typography'
 
 import { toggleValueInList } from '../../../../utils'
 import { useSettingsContext } from '../../../../settings'
-import useOperators from '../../../../logic/useOperators'
+import { useOperators } from '../../../../logic'
 import { ALL_ACTIVE_OPERATOR_IDS } from '../../../../constants'
 
 import './styles.scss'
