+++ src/dashboards/Chrono/index.tsx
@@ -9,7 +9,7 @@ import { FormFactor } from '@entur/sdk/lib/mobility/types'
 import {
     useBikeRentalStations,
     useStopPlacesWithDepartures,
-    useScooters,
+    useMobility,
     useWalkInfo,
 } from '../../logic'
 import DashboardWrapper from '../../containers/DashboardWrapper'
@@ -103,7 +103,7 @@ const ChronoDashboard = ({ history }: Props): JSX.Element | null => {
 
     const bikeRentalStations = useBikeRentalStations()
     let stopPlacesWithDepartures = useStopPlacesWithDepartures()
-    const scooters = useScooters([FormFactor.SCOOTER])
+    const scooters = useMobility([FormFactor.SCOOTER])
 
     const hasData = Boolean(
         bikeRentalStations?.length ||
